backend/requirements.txt
Django>=5.0
djangorestframework
djangorestframework-simplejwt
django-cors-headers
celery
redis
psycopg2-binary
python-dotenv
